NP-1292 Use hook for authentication
