+++ src/pages/user/authority/AuthorityModal.tsx
@@ -1,9 +1,8 @@
 import React, { FC, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { Button } from '@material-ui/core';
 import styled from 'styled-components';
-import { Auth } from 'aws-amplify';
 
 import { Authority } from '../../../types/authority.types';
 import { RootStore } from '../../../redux/reducers/rootReducer';
@@ -13,8 +12,7 @@ import Modal from '../../../components/Modal';
 import { StyledRightAlignedButtonWrapper } from '../../../components/styled/Wrappers';
 import NormalText from '../../../components/NormalText';
 import ConfirmDialog from '../../../components/ConfirmDialog';
-import { USE_MOCK_DATA } from '../../../utils/constants';
-import { logoutSuccess } from '../../../redux/actions/authActions';
+import { useAuthentication } from '../../../utils/hooks/useAuthentication';
 
 const StyledButtonContainer = styled(StyledRightAlignedButtonWrapper)`
   margin-top: 2rem;
@@ -37,16 +35,7 @@ const AuthorityModal: FC<AuthorityModalProps> = ({ handleNextClick }) => {
   const { t } = useTranslation('profile');
   const { authority } = useSelector((store: RootStore) => store.user);
   const [openCancelConfirmation, setOpenCancelConfirmation] = useState(false);
-
-  const dispatch = useDispatch();
-  const logOut = () => {
-    if (USE_MOCK_DATA) {
-      dispatch(logoutSuccess());
-      window.location.pathname = '/logout';
-    } else {
-      Auth.signOut();
-    }
-  };
+  const { handleLogout } = useAuthentication();
 
   const toggleCancelConfirmation = () => setOpenCancelConfirmation((state) => !state);
 
@@ -90,7 +79,7 @@ const AuthorityModal: FC<AuthorityModalProps> = ({ handleNextClick }) => {
       <ConfirmDialog
         open={openCancelConfirmation}
         title={t('authority.cancel_connect_authority_title')}
-        onAccept={logOut}
+        onAccept={handleLogout}
         onCancel={toggleCancelConfirmation}>
         {t('authority.cancel_connect_authority_text')}
       </ConfirmDialog>
